Add support for "legacy" MiLight / easybulb devices.
These are the original generation MiLight bulbs that use a four-bulb
wifi hub (not the newer "v6" bulbs"). There's only support for the RGBW
bulbs at the moment - each of the RGB/W/RGBW bulbs have a different
instruction set which is documented at
https://github.com/BKrajancic/LimitlessLED-DevAPI/blob/master/LimitlessLed_Dev_Markdown.md
(the bottom half of the page talks about these "v2" bulbs).